requirements.txt
sendgrid==6.11.0
jinja2==3.1.4
python-dotenv==1.0.1